projections/requirements.txt
# Python dependencies for pension projections
supabase>=2.0.0
python-dotenv>=1.0.0

